# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holivita AI API",
        "description": "API для онбординг-чата с AI-ассистентом, суммаризации и голосового ввода (OpenAI Whisper). Авторизация отключена для MVP — user_id передаётся вручную. Модули: Onboarding (синхронный и асинхронный чат), Summaries (суммаризации сессий), Voice (транскрипция аудио). Формат ответов: успех — {success: true, data: {...}}, ошибка — {success: false, error: \"...\"}.",
        "contact": {
            "name": "Holivita AI Team",
            "email": "support@holivita.ai"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/onboarding/validate-user": {
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Валидация пользователя перед началом онбординга",
                "description": "Проверяет, может ли пользователь начать новую сессию онбординга. Если у пользователя уже есть активная сессия — возвращается конфликт (409) с ID существующей сессии. Используйте этот эндпоинт перед вызовом /onboarding/chat для проверки доступности.",
                "operationId": "validateUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь валиден, можно начинать онбординг",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User ID валиден"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "У пользователя уже есть активная сессия",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "У пользователя уже есть активная сессия."
                                        },
                                        "active_session_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации входных данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User ID обязателен."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/onboarding/chat": {
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Отправка сообщения в синхронный чат онбординга",
                "description": "Отправляет сообщение пользователя AI-ассистенту и возвращает ответ синхронно. Если message пустое или не передано — инициирует начало онбординга (приветственное сообщение от ассистента). Если у пользователя нет активной сессии — она создаётся автоматически.",
                "operationId": "onboardingChat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    },
                                    "session_id": {
                                        "description": "ID сессии (опционально, создаётся автоматически)",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "message": {
                                        "description": "Текст сообщения. Если пустое — запускается приветствие ассистента.",
                                        "type": "string",
                                        "maxLength": 2000,
                                        "example": "Меня зовут Иван, мне 35 лет",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ ассистента получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Здравствуйте! Расскажите о ваших целях здоровья."
                                                },
                                                "completed": {
                                                    "description": "true, если ассистент считает онбординг завершённым",
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "session_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации входных данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User ID обязателен."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "AI-сервис временно недоступен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Не удалось получить ответ от ассистента. Попробуйте позже."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}